remuxer: Add an option to merge last fragments if they brach a given threshold
This is useful, for example, if you have a 6.2 second file with a 3 second
minimum fragment duration, in that it can merged the last to fragments into
one 3.2 second fragment instead of a 3 and 0.2 second segment.

The CLI argument takes a percentage of the minimum fragment duration. For
example, with a minimum fragment duration of 3, and a threshold of 0.5, the
last two segments will be merged if, when combined, they would be less than
or equal to 4.5 seconds long.

Signed-off-by: Derek Buitenhuis <derek.buitenhuis@gmail.com>