hevc: Fix Dolby Vision level pixels-per-second checks
These values are max values, and thus are valid for any given level,
meaning we should be using <= and not < to check.

This fixes the case, where, for example, a 3840x2160 video at exactly
integer 60 fps gets incorrectly set as level 10, when it should be 9,
causing playback checks to fail in Safari.

As per section 2.2 here in:
    dolby-vision-streams-within-the-http-live-streaming-format-v2.0-13-november-2018.pdf

Signed-off-by: Derek Buitenhuis <derek.buitenhuis@gmail.com>